CPU: Move flags to `flags.rs` file
