+++ emu/src/cpu/mod.rs
@@ -3,6 +3,7 @@ pub mod arm7tdmi;
 mod condition;
 mod cpu_modes;
 mod data_processing;
+mod flags;
 mod instruction;
 mod move_compare_add_sub;
 mod opcode;
